Add keystring retrieval to ADX cracker.
There is now an option to specify a file to search for key strings in. This would usually be the game's main executable.
When a key is found, the key will be checked to see if it matches any of the strings found in the executable, and prints any that match.